+++ api/RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## in develop
 
+* Adds `DxApi.addTags` method
+* Fixes `DxFindDataObjects` when used with `tags` constraint
 * Handles record results in `DxFindDataObjects`
 
 ## 0.12.0 (2021-11-18)